controller/sqm.lua: Reflect Luci changes and relocate/rename file
Reflect new Luci pages and locations for SQM overview and details. Factor
out common app description strings used on other pages, and clean up
Lua requires. Create an empty sqm config file if none is found, rather
than abort silently. Relocate the file to luasrc/controller/sqm.lua to
better align with the openwrt-luci repo standard.
